backend/requirements.txt
uvicorn
python-dotenv
fastapi
pydantic
requests